+++ setup.py
@@ -4,8 +4,8 @@ setup(
     name="videoswin",
     packages=find_packages(exclude=["notebooks", "assets"]),
     version="1.0.0",
-    license="MIT",
-    description="Video Swin Transformerss in Keras",
+    license="Apache License 2.0",
+    description="Video Swin Transformerss in Keras 3",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
     author="Mohammed Innat",
@@ -15,16 +15,31 @@ setup(
     install_requires=[
         "opencv-python>=4.1.2",
         "keras-nightly",
+        "tensorflow-datasets",
     ],
-    setup_requires=[
-        "pytest-runner",
-    ],
-    tests_require=["pytest"],
+    extras_require={
+        "tests": [
+            "flake8",
+            "isort",
+            "black[jupyter]",
+            "pytest",
+            "pycocotools",
+        ],
+        "examples": ["tensorflow_datasets", "matplotlib"],
+    },
+    python_requires=">=3.9",
     classifiers=[
-        "Development Status :: 1 - Planning",
-        "Intended Audience :: Developers",
-        "Topic :: Scientific/Engineering :: Artificial Intelligence",
-        "License :: OSI Approved :: MIT License",
+        "Development Status :: 3 - Alpha",
+        "Programming Language :: Python :: 3",
         "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3 :: Only",
+        "Operating System :: Unix",
+        "Operating System :: Microsoft :: Windows",
+        "Operating System :: MacOS",
+        "Intended Audience :: Science/Research",
+        "Topic :: Scientific/Engineering",
+        "Topic :: Software Development",
     ],
 )
